+++ changelog.md
@@ -52,6 +52,7 @@
 
 * Removed all deprecated methods and properties from `Kymo`, `Scan`, `PointScan`, and `ImageStack`.
 * Removed deprecated argument `exclude` from `Kymo.line_timestamp_ranges()` and `Scan.frame_timestamp_ranges()`.
+* Removed deprecated argument `line_width` from `lk.track_greedy()`.
 
 ## v0.13.3 | 2023-01-26